+++ src/chemflow_client/__init__.py
@@ -1,6 +1,7 @@
 """Public ChemFlow Python client."""
 
 from .client import ChemFlow3DClient, chat3d
+from .constants import DEFAULT_BASE_URL
 from .exceptions import (
     ChemFlowConfigurationError,
     ChemFlowError,
@@ -14,6 +15,7 @@ __all__ = [
     "chat3d",
     "ChemFlow3DClient",
     "Chat3DWidget",
+    "DEFAULT_BASE_URL",
     "ChemFlowError",
     "ChemFlowConfigurationError",
     "ChemFlowHttpError",