+++ src/chemflow_client/constants.py
@@ -1,0 +1,4 @@
+"""Public constants for chemflow-client."""
+
+DEFAULT_BASE_URL = "https://chemcloud.info"
+